added contract address to view page
