+++ src/main/tile-nft-frontend/src/view/loadViewTokenData.js
@@ -1,4 +1,5 @@
 import loadBlockExplorerUrl from "./loadBlockExplorerUrl";
+import loadContractAddress from "./loadContractAddress";
 
 const noop = () => {};
 
@@ -11,31 +12,6 @@ const loadViewTokenData = ({
       return `${window.location.origin}/api/image/tile/get/${tokenId}`;
     }
     return getMetadataToUpdateTokenUrl(tokenId, metadataToUpdate);
-  }
-
-  const loadContractAddress = () => {
-    return fetch(
-        `${window.location.origin}/api/contract/getAddress`,
-        {method: 'get'}
-    )
-    .then(response => {
-      if (response.status === 200) {
-        return response.json();
-      }
-      return null;
-    })
-    .then(json => {
-      console.log("ViewToken loadContractAddress json found", json);
-      if (json === null) {
-        let errorMessage = 'loadContractAddress json is null';
-        console.log(errorMessage);
-        throw errorMessage;
-      }
-      return json
-    })
-    .catch(err => {
-      console.log("Error caught!!!", err);
-    });
   }
 
   const loadTokenImageData = (tokenId) => {
